backend/detection-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
ultralytics==8.1.4
torch==2.1.2
torchvision==0.16.2
scikit-image==0.22.0
exifread==3.0.0
python-dotenv==1.0.0
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
